删除文章分类关联 删除分类数据库 删除页面数据库 修改文章数据库：删除分类字段、增加view字段 修改删除对应模型、路由、控制器

+++ database/migrations/2020_02_25_105923_create_posts_table.php
@@ -16,8 +16,8 @@ class CreatePostsTable extends Migration
     {
         Schema::create('posts', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->integer('author_id')->default(0)->comment('作者ID');
-            $table->integer('category_id')->default(0)->comment('分类ID');
+            $table->bigInteger('author_id')->default(0)->comment('作者ID');
+            $table->bigInteger('category_id')->default(0)->comment('分类ID');
             $table->string('title')->default('')->comment('文章标题');
             $table->string('image')->nullable()->comment('文章封面');
             $table->text('excerpt')->comment('文章摘要');
